+++ tests/unit/eyes_core/test_visual_locator.py
@@ -20,13 +20,13 @@ def test_wrong_visual_locator_names():
 @pytest.mark.parametrize("name", ["Some name", "Create"])
 def test_visual_locator_name(name):
     vl = VisualLocator.name(name)
-    assert vl.names == [name]
-    assert vl.is_first_only == True
+    assert vl.values.names == [name]
+    assert vl.values.is_first_only == True
     vl = vl.all()
-    assert vl.is_first_only == False
+    assert vl.values.is_first_only == False
 
     vl = VisualLocator.names(name, name)
-    assert vl.names == [name, name]
+    assert vl.values.names == [name, name]
 
     vl = VisualLocator.names([name, name])
-    assert vl.names == [name, name]
+    assert vl.values.names == [name, name]